Add overlays for VAST
